Update prog1/week 2/module 6/string.c

+++ prog1/week 2/module 6/string.c
@@ -34,7 +34,7 @@
         
         for (int i = 0; i < strlen(input); i++)
         {
-            if (input[i] = 'E' || input[i] = 'e')
+            if ((int)input[i] = 'E' || input[i] = 'e')
             {
                 n++;
             }
